refactor: reduce repetitive code
